fixed custom nav selecting bug
