+++ src/Components/App.tsx
@@ -11,7 +11,6 @@ const Gear = lazy(() => import("./Gear"));
 const Main = lazy(() => import("./Main"));
 const FallBack = lazy(() => import("./FallBack"));
 const ProjectsPage = lazy(() => import("./ProjectsPage"));
-import { getRelativeURL } from "../helpers/getRelativeURL";
 import "../css/app.css";
 
 const defaultThemeContextValue: ThemeContextType = {
@@ -52,40 +51,6 @@ function App() {
 		html.style.backgroundColor = "#FFFFFF";
 		body.style.backgroundColor = "#FFFFFF";
 		body.style.color = "#000000";
-	}
-
-	// Get all sections that have an ID defined
-	const sections = document.querySelectorAll("section[id]");
-
-	// Add an event listener listening for scroll
-	var currentURL = getRelativeURL();
-	if (currentURL === "/") {
-		window.addEventListener("scroll", navHighlighter);
-	}
-
-	function navHighlighter() {
-		// Get current scroll position
-		let scrollY = window.scrollY;
-
-		// Now we loop through sections to get height, top and ID values for each
-		sections.forEach((current: any) => {
-			const sectionHeight = current.offsetHeight;
-			const sectionTop = current.offsetTop - 400; // earlier I was substracting 62
-			var sectionId = current.getAttribute("id");
-			/*
-			- If our current scroll position enters the space where current section on screen is, add .active class to corresponding navigation link, else remove it
-			- To know which link needs an active class, we use sectionId variable we are getting while looping through sections as an selector
-			*/
-			if (scrollY > sectionTop && scrollY <= sectionTop + sectionHeight) {
-				document
-					.querySelector(".custom-nav a[href*=" + sectionId + "]")
-					?.classList.add("active");
-			} else {
-				document
-					.querySelector(".custom-nav a[href*=" + sectionId + "]")
-					?.classList.remove("active");
-			}
-		});
 	}
 	return (
 		<ThemeContext.Provider value={{ themeflag, theme, toggleTheme }}>
